test/web-platform-tests/to-upstream/cssom-view/Element-clientWidth.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Element.prototype.clientWidth</title>
<link rel=help href="https://drafts.csswg.org/cssom-view/#dom-element-clientwidth">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  assert_equals(el.clientWidth, 0);

}, "clientWidth should return 0 for an element with no associated CSS layout box");
</script>
